Separate individual project card as its own component. Delete index-copy.js
